+++ init.rb
@@ -5,8 +5,10 @@ require_dependency File.expand_path('../lib/redmine_external_issue_links/source_
 Redmine::Plugin.register :redmine_external_issue_links do
   name 'Redmine External Source Links'
   author 'Konovalov Semyon'
-  description 'Adds external source links to Redmine issues with icons, ordering, journal notes and REST API.'
-  version '1.0.2'
+  author_url 'https://github.com/SKonovalovS'
+  description 'Adds external source links to Redmine issues with icons, ordering, journal notes and REST API. Plugin: https://github.com/SKonovalovS/redmine-external-source'
+  url 'https://github.com/SKonovalovS/redmine-external-source'
+  version '1.0.3'
   requires_redmine version_or_higher: '5.0.0'
 
   project_module :external_issue_links do